resources/tools/customs/youtube-search/requirements.txt
google-api-python-client>=2.100.0
youtube-transcript-api>=1.0.0
python-dotenv>=1.0.0
